feat:add logic for saveRecipes

+++ backend/src/main/java/salt/tobias/meal/recipe/controller/RecipeRequestDto.java
@@ -2,6 +2,6 @@ package salt.tobias.meal.recipe.controller;
 
 import salt.tobias.meal.recipeApi.model.Recipe;
 
-public record RecipeRequestDto(double[] servings, Recipe[] recipes) {
+public record RecipeRequestDto(long[] servings, Recipe[] recipes) {
 
 }
